Added Options module
