+++ test/runner_test.exs
@@ -11,15 +11,6 @@ defmodule RunnerTest do
     assert capture_io(response) == "Okay, Goodbye!\n"
   end
 
-  test "user can see the player's markers" do
-    response = fn ->
-      assert Runner.show_markers({"X", "O"}) == {"X", "O"}
-    end
-
-    assert capture_io(response) ==
-             "Let's begin!\n\nPlayer One - Your marker is \"X\".\nPlayer Two - Your marker is \"O\".\n\n"
-  end
-
   test "user can be prompted to choose again" do
     response = fn ->
       assert Runner.choose_again() == :ok
@@ -57,20 +48,6 @@ defmodule RunnerTest do
              "This game is a draw.\n"
   end
 
-  test "user can enter Y to play a single player game" do
-    output = fn -> Runner.single_player_game() end
-
-    assert capture_io([input: "Y\nY", capture_prompt: false], output) =~
-             "Let's begin!\n\nPlayer One - Your marker is \"X\".\nPlayer Two - Your marker is \"O\".\n\n"
-  end
-
-  test "user can only enter Y/N to play a single player game" do
-    output = fn -> Runner.single_player_game("*") end
-
-    assert capture_io([input: "@\nY\nY", capture_prompt: false], output) =~
-             "Please enter Y/N only\n"
-  end
-
   test "user can play a single player game" do
     output = fn -> Runner.start_game() end
 
@@ -79,46 +56,5 @@ defmodule RunnerTest do
              output
            ) =~
              "You have started a new game of TTT!\n"
-  end
-
-  test "user can choose default markers" do
-    response = fn ->
-      assert Runner.use_default_markers("Y") == {"X", "O"}
-    end
-
-    assert capture_io(response) ==
-             "Let's begin!\n\nPlayer One - Your marker is \"X\".\nPlayer Two - Your marker is \"O\".\n\n"
-  end
-
-  test "user can choose custom markers" do
-    response = fn -> Runner.use_default_markers("N") end
-
-    assert capture_io([input: "+\n#"], response) =~
-             "Please enter a marker for Player One:\n"
-  end
-
-  test "user cannot enter anything other than Y/N when prompted to use default markers" do
-    response = fn -> Runner.use_default_markers("%") end
-
-    assert capture_io([input: "y"], response) =~
-             "Please enter Y/N only\n"
-  end
-
-  test "when get_markers is passed a custom marker it gets the second marker" do
-    output = fn -> Runner.get_markers("+") end
-
-    assert capture_io([input: "@", capture_prompt: false], output) ==
-             "Let's begin!\n\nPlayer One - Your marker is \"+\".\nPlayer Two - Your marker is \"@\".\n\n"
-  end
-
-  test "users cannot choose two of the same custom markers" do
-    output = fn -> Runner.different_markers?({"+", "+"}) end
-
-    assert capture_io([input: "@", capture_prompt: false], output) ==
-             "Choose a different symbol - they cannot be the same!\nLet's begin!\n\nPlayer One - Your marker is \"+\".\nPlayer Two - Your marker is \"@\".\n\n"
-  end
-
-  test "different_markers? returns the markers when they are different" do
-    assert Runner.different_markers?({"+", "^"}) == {"+", "^"}
   end
 end
